requirements.txt
git+https://github.com/Rapptz/discord-ext-menus@84caae8038d0d
discord.py
asyncpraw
BotUtils
lxml
psutil
asyncpg
click
parsedatetime
lru_dict
python-dateutil
uvloop
dateparser
pandas
asyncprawcore
pytz
urllib3
aiohttp
discord
matplotlib
dataframe_image
celery[redis]
colorlog
SQLAlchemy
pylibmc